fixed the pick date dialogs
