+++ app/src/main/java/com/binus/pekalongancityguide/Layout/AddItinerary.java
@@ -1,7 +1,5 @@
 package com.binus.pekalongancityguide.Layout;
 
-import android.app.AlertDialog;
-import android.app.DatePickerDialog;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -22,35 +20,25 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.binus.pekalongancityguide.Adapter.IterAdapter;
 import com.binus.pekalongancityguide.ItemTemplate.Destination;
 import com.binus.pekalongancityguide.R;
-import com.binus.pekalongancityguide.databinding.DialogChooseDateBinding;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Collections;
-import java.util.Date;
-import java.util.Locale;
 
 import static android.content.ContentValues.TAG;
 
 public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClickListener {
-    private String categoryId, category, startDate, endDate;
+    private String categoryId, category, startDate, endDate, uid;
     public IterAdapter iterAdapter;
     private RecyclerView iterRV;
-    private Button addIter, addDate;
+    private Button addIter;
     private RelativeLayout selectLayout;
     private TextView selectTv;
-    private ImageButton selectCancel, startBtn, endBtn;
-    ;
-    private EditText startEt, endEt;
-    private Calendar calendar;
-    private int startDay, startMonth, startYear, endDay, endMonth, endYear, counter;
+    private ImageButton selectCancel;
     public ArrayList<Destination> destinationArrayList, selectedItems;
     private View view;
     private ItineraryPager itineraryPager;
@@ -59,12 +47,15 @@ public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClic
     public AddItinerary() {
     }
 
-    public static AddItinerary newInstance(String categoryId, String category, String uid) {
+    public static AddItinerary newInstance(String categoryId, String categoryName, String categoryUid, String startDate, String endDate) {
         AddItinerary fragment = new AddItinerary();
         Bundle args = new Bundle();
         args.putString("categoryId", categoryId);
-        args.putString("category", category);
-        args.putString("uid", uid);
+        args.putString("category", categoryName);
+        args.putString("uid", categoryUid);
+        args.putString("startDate", startDate);
+        args.putString("endDate", endDate);
+        Log.d(TAG, "newInstance: categoryId=" + categoryId + ", categoryName=" + categoryName + ", categoryUid=" + categoryUid + ", startDate=" + startDate + ", endDate=" + endDate);
         fragment.setArguments(args);
         return fragment;
     }
@@ -79,6 +70,10 @@ public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClic
         if (getArguments() != null) {
             categoryId = getArguments().getString("categoryId");
             category = getArguments().getString("category");
+            uid = getArguments().getString("uid");
+            startDate = getArguments().getString("startDate");
+            endDate = getArguments().getString("endDate");
+            Log.d(TAG, "onCreate: categoryId=" + categoryId + ", category=" + category + ", uid=" + uid + ", startDate=" + startDate + ", endDate=" + endDate);
         }
     }
 
@@ -87,7 +82,6 @@ public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClic
                              Bundle savedInstanceState) {
         view = inflater.inflate(R.layout.fragment_add_itinerary, container, false);
         init();
-        showPickDateDialog();
         checkSelect();
         EditText iterSearch = view.findViewById(R.id.search_iter);
         iterSearch.addTextChangedListener(new TextWatcher() {
@@ -129,98 +123,6 @@ public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClic
         selectLayout = view.findViewById(R.id.select_layout);
         selectCancel = view.findViewById(R.id.select_cancel);
     }
-
-    private void showPickDateDialog() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        DialogChooseDateBinding chooseDateBinding = DialogChooseDateBinding.inflate(getLayoutInflater());
-        builder.setView(chooseDateBinding.getRoot());
-        startEt = chooseDateBinding.startDateEt;
-        endEt = chooseDateBinding.endDateEt;
-        startBtn = chooseDateBinding.startPickerBtn;
-        endBtn = chooseDateBinding.endPickerBtn;
-        addDate = chooseDateBinding.addDateBtn;
-
-        calendar = Calendar.getInstance();
-
-        startBtn.setOnClickListener(v -> showStartCalendar());
-
-        startEt.setOnClickListener(v -> showStartCalendar());
-
-        endBtn.setOnClickListener(v -> showEndCalendar());
-
-        endEt.setOnClickListener(v -> showEndCalendar());
-        AlertDialog dialog = builder.create();
-        dialog.getWindow().setBackgroundDrawableResource(R.drawable.dialog_background);
-        dialog.show();
-        addDate.setOnClickListener(v -> dialog.dismiss());
-    }
-
-    private void showStartCalendar() {
-        startYear = calendar.get(Calendar.YEAR);
-        startMonth = calendar.get(Calendar.MONTH);
-        startDay = calendar.get(Calendar.DAY_OF_MONTH);
-
-        DatePickerDialog dialog = new DatePickerDialog(getContext(), (dateView, year, month, dayOfMonth) -> {
-            startYear = year;
-            startMonth = month;
-            startDay = dayOfMonth;
-            SimpleDateFormat format = new SimpleDateFormat("EEEE, dd MMMM yyyy", Locale.getDefault());
-            startDate = format.format(new Date(startYear - 1900, startMonth, startDay));
-            startEt.setText(startDate);
-        }, startYear, startMonth, startDay);
-
-        dialog.getDatePicker().setMinDate(calendar.getTimeInMillis());
-
-        if (endDate != null) {
-            SimpleDateFormat format = new SimpleDateFormat("EEEE, dd MMMM yyyy", Locale.getDefault());
-            try {
-                Date endDateObj = format.parse(endDate);
-                dialog.getDatePicker().setMaxDate(endDateObj.getTime());
-            } catch (ParseException e) {
-                e.printStackTrace();
-            }
-        }
-
-        dialog.getWindow().setBackgroundDrawableResource(R.color.palette_4);
-        dialog.show();
-    }
-
-    private void showEndCalendar() {
-        endYear = calendar.get(Calendar.YEAR);
-        endMonth = calendar.get(Calendar.MONTH);
-        endDay = calendar.get(Calendar.DAY_OF_MONTH);
-
-        DatePickerDialog dialog = new DatePickerDialog(getContext(), (dateView, year, month, dayOfMonth) -> {
-            endYear = year;
-            endMonth = month;
-            endDay = dayOfMonth;
-            SimpleDateFormat format = new SimpleDateFormat("EEEE, dd MMMM yyyy", Locale.getDefault());
-            endDate = format.format(new Date(endYear - 1900, endMonth, endDay));
-            endEt.setText(endDate);
-        }, endYear, endMonth, endDay);
-
-        Calendar minDateCalendar = Calendar.getInstance();
-        if (startDate != null) {
-            SimpleDateFormat format = new SimpleDateFormat("EEEE, dd MMMM yyyy", Locale.getDefault());
-            try {
-                Date startDateObj = format.parse(startDate);
-                minDateCalendar.setTime(startDateObj);
-            } catch (ParseException e) {
-                e.printStackTrace();
-            }
-        } else {
-            minDateCalendar.setTime(calendar.getTime());
-        }
-        if (minDateCalendar.after(calendar)) {
-            dialog.getDatePicker().setMinDate(minDateCalendar.getTimeInMillis());
-        } else {
-            dialog.getDatePicker().setMinDate(calendar.getTimeInMillis());
-        }
-
-        dialog.getWindow().setBackgroundDrawableResource(R.color.palette_4);
-        dialog.show();
-    }
-
     private void loadDestinations() {
         destinationArrayList = new ArrayList<>();
         DatabaseReference reference = FirebaseDatabase.getInstance("https://pekalongan-city-guide-5bf2e-default-rtdb.asia-southeast1.firebasedatabase.app/").getReference("Destination");
@@ -286,6 +188,7 @@ public class AddItinerary extends Fragment implements IterAdapter.OnItemLongClic
     public void checkSelect() {
         if (iterAdapter != null) {
             selectedItems = iterAdapter.getSelectedItems();
+            int counter;
             if (selectedItems.size() < 1) {
                 selectLayout.setVisibility(View.GONE);
                 addIter.setVisibility(View.INVISIBLE);
